Get App into working state with read only fsystem
This commit improves the previous solution to transfer the app
into a working state.
For this the commit creates two temporary volumes which will then
be mounted to the path nginx tries to write to. This way the
filesystem can stay read only and the app will still work.
Having the filesystem read only can improve security and is
therefore the better solution here.