Increase replica count to 5
This commit will increase the replica count from the default 2
to 5.
It also adjusts the resources of the container, because the
default values would cause the minikube docker container to run
out of cpu.
Considering that it has 2 cores=2000m it seems sensible to divide
these between 5 pods which would leave the limit at 400m, and
setting the requested just a bit below to 250 millicpu.