Add liveness probe
This commit adds a liveness http request on port 80 to ensure
that the pod is still running and (in this case) nginx is still
running.
The specific parameters would need to be discussed and depend
on the requirements of the application. For the purpose of this
task I decided to set the initial delay seconds to 10 to give
the app plenty of time to start before the first request is
executed. Period seconds was set to 10 to ensure that the request
happens every 10 seconds. Timeout seconds defines how long we
allow the request to take, and if its not received in 5 seconds,
it will be considered a failure. The failure threshold is set to
1 as there is no good reason why this app should exceed this.
This was tested by accessing one of the pods, stopping nginx and
waiting for the pod to be recreated, which worked fine.