Reorganise project into separate component files
This commit should reorganise the project and provide a final
version of the solution for the test.
It does not only split the kubernetes manifest in multiple ones
using different descriptive charts, it also reorganises the python
code into multiple separate files.
The project is now structured in a way that each kubernetes
component can be managed in a separate python file. Creating a
separate python file for volumes was a conscious decision to
follow this structure. It could also be argued that it belongs in
the deployment file to have a structure separated by functionality.
In addition to that I added a solution part to the Readme to inform
the user how to run the final version of this project.